Push Bai tap

+++ Module 2/Baitap/src/ss12_SearchAlgorithm/thuchanh/AlgorithmComplexityTest.java
@@ -9,7 +9,7 @@ public class AlgorithmComplexityTest {
         String str = scanner.nextLine();
         int[] frequentChar = new int[255];
         for (int i = 0; i < str.length(); i++) {
-            int ascii = (int) str.charAt(i);
+            int ascii = str.charAt(i);
             frequentChar[ascii] += 1;
         }
         int max = 0 ;
